wip(mika#1450): rescued impl (dispatch-lib recovery) (#1492)
* docs(plans): groom mika issue#1450 (content-only, architect pending)

* wip(mika#1450): impl staged by post-flight recovery (mika#1282)

Content written by pilot session 2113e242-c228-4b84-8b9a-6ec158a1cb15 but git commit was never invoked.
Auto-rescued by dispatch-lib dirty-worktree detection.
Scaffold paths excluded (mika#1288, mika#1419).

---------
