+++ crates/mika-agent/src/agent.rs
@@ -24,22 +24,19 @@ use crate::mcp::McpManager;
 use crate::messaging::MessageSender;
 use crate::post_condition::{GuardDecision, POST_CONDITION_GUARDS};
 use crate::prompt;
-use crate::secret_scrubber::scrub_secrets;
 use crate::skills::SkillRegistry;
-use crate::skills::builtin_handlers;
 use crate::skills::context;
 use crate::skills::executor;
 use crate::skills::index::{ResolvedSkillTool, SkillEntry};
-use crate::skills::manifest::ToolHandler;
 use crate::skills::matcher::{MatchReason, MatchedSkill};
 use crate::skills::review_filter;
-use crate::tools::{SkillPathInfo, ToolContext, ToolOutput, ToolRegistry};
+use crate::tool_execution::{
+    ToolCallSummary, format_tool_summary_block, process_tool_calls, tool_calls_metadata_json,
+};
+use crate::tools::{SkillPathInfo, ToolContext, ToolRegistry};
 use mika_common::config::Settings;
 use mika_common::embedding::EmbeddingClient;
 use mika_common::llm::ProviderKind;
-
-/// Max chars of serialized tool input to include in timeout log lines (#900).
-const TOOL_TIMEOUT_INPUT_EXCERPT_LEN: usize = 200;
 
 /// Skills whose LLM output legitimately contains `Verdict:` lines.
 /// Used by the dev-groom fabrication guard (#1133, #1254) to exempt
@@ -262,160 +259,6 @@ impl LoopMode {
             Self::Team => "team agent",
         }
     }
-}
-
-/// Summary of a single tool call for persistence in conversation metadata.
-#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct ToolCallSummary {
-    pub step: u32,
-    pub name: String,
-    pub input_summary: String,
-    pub output_summary: String,
-    pub success: bool,
-    /// True when the tool output starts with a non-zero exit code prefix
-    /// (e.g. "Exit code: 1" or "Killed by signal: 9"). When set, `success`
-    /// is `false`. This field provides additional detail about *why* it failed.
-    #[serde(default)]
-    pub non_zero_exit: bool,
-}
-
-/// Check whether tool output content starts with a non-zero exit code prefix
-/// produced by the exec handler for subprocesses that exit non-zero.
-fn has_non_zero_exit_prefix(content: &str) -> bool {
-    if let Some(rest) = content.strip_prefix("Exit code: ") {
-        // "Exit code: 0" is never emitted (exit 0 has no prefix), but guard anyway
-        rest.starts_with(|c: char| c.is_ascii_digit()) && !rest.starts_with('0')
-    } else {
-        content.starts_with("Killed by signal:")
-    }
-}
-
-/// Truncate a string to approximately `max_len` bytes, appending "..." if truncated.
-/// Always cuts at a valid UTF-8 char boundary to avoid panics on multi-byte input.
-fn truncate_summary(s: &str, max_len: usize) -> String {
-    if s.len() <= max_len {
-        s.to_string()
-    } else {
-        let cut = max_len.saturating_sub(3);
-        // Walk back to a valid char boundary
-        let mut boundary = cut;
-        while boundary > 0 && !s.is_char_boundary(boundary) {
-            boundary -= 1;
-        }
-        format!("{}...", &s[..boundary])
-    }
-}
-
-/// Serialize tool call summaries to JSON metadata string, capped at [`crate::planning::policy::TOOL_METADATA_MAX`].
-///
-/// Strategy: preserve all entries by progressively truncating per-field content.
-/// 1. Try full serialization with the initial field lengths.
-/// 2. If over budget, re-truncate `input_summary` and `output_summary` to fit all entries.
-/// 3. Only as a last resort, drop tail entries (with a warning).
-pub fn tool_calls_metadata_json(summaries: &[ToolCallSummary]) -> Option<String> {
-    if summaries.is_empty() {
-        return None;
-    }
-    let wrapper = serde_json::json!({ "tool_calls": summaries });
-    let json = serde_json::to_string(&wrapper).ok()?;
-    if json.len() <= crate::planning::policy::TOOL_METADATA_MAX {
-        return Some(json);
-    }
-
-    // Phase 1: Aggressively re-truncate fields to fit all entries.
-    let shrunk: Vec<ToolCallSummary> = summaries
-        .iter()
-        .map(|s| ToolCallSummary {
-            step: s.step,
-            name: s.name.clone(),
-            input_summary: truncate_summary(&s.input_summary, 30),
-            output_summary: truncate_summary(&s.output_summary, 50),
-            success: s.success,
-            non_zero_exit: s.non_zero_exit,
-        })
-        .collect();
-    let wrapper = serde_json::json!({ "tool_calls": shrunk });
-    if let Ok(json) = serde_json::to_string(&wrapper)
-        && json.len() <= crate::planning::policy::TOOL_METADATA_MAX
-    {
-        return Some(json);
-    }
-
-    // Phase 2: Last resort — drop tail entries from the already-shrunk vector.
-    warn!(
-        total_entries = summaries.len(),
-        max = crate::planning::policy::TOOL_METADATA_MAX,
-        "tool_calls metadata exceeds cap after field truncation, dropping tail entries"
-    );
-    for count in (1..shrunk.len()).rev() {
-        let wrapper = serde_json::json!({ "tool_calls": &shrunk[..count] });
-        if let Ok(json) = serde_json::to_string(&wrapper)
-            && json.len() <= crate::planning::policy::TOOL_METADATA_MAX
-        {
-            return Some(json);
-        }
-    }
-    warn!(
-        total_entries = summaries.len(),
-        "tool_calls metadata: unable to fit even a single entry, returning None"
-    );
-    None
-}
-
-/// Format tool call metadata into a concise summary block for injection into history.
-///
-/// Includes truncated input so the agent can introspect what arguments it passed
-/// (e.g., "what command did you send?") and output for result context.
-/// Malformed entries are skipped rather than causing the entire block to be dropped.
-pub fn format_tool_summary_block(metadata_json: &str) -> Option<String> {
-    let parsed: serde_json::Value = serde_json::from_str(metadata_json).ok()?;
-    let calls = parsed.get("tool_calls")?.as_array()?;
-    if calls.is_empty() {
-        return None;
-    }
-    let parts: Vec<String> = calls
-        .iter()
-        .filter_map(|call| {
-            let name = call.get("name")?.as_str()?;
-            let input = call
-                .get("input_summary")
-                .and_then(|v| v.as_str())
-                .unwrap_or("");
-            let output = call
-                .get("output_summary")
-                .and_then(|v| v.as_str())
-                .unwrap_or("");
-            let success = call
-                .get("success")
-                .and_then(|v| v.as_bool())
-                .unwrap_or(true);
-            let non_zero_exit = call
-                .get("non_zero_exit")
-                .and_then(|v| v.as_bool())
-                .unwrap_or(false);
-            let status = if non_zero_exit {
-                " [NON-ZERO]"
-            } else if !success {
-                " [FAILED]"
-            } else {
-                ""
-            };
-            let short_input = truncate_summary(input, 60);
-            let short_output = truncate_summary(output, 80);
-            if short_input.is_empty() {
-                Some(format!("{name}{status} → {short_output}"))
-            } else {
-                Some(format!("{name}({short_input}){status} → {short_output}"))
-            }
-        })
-        .collect();
-    if parts.is_empty() {
-        return None;
-    }
-    Some(format!(
-        "\n<context type=\"tool_history\" trust=\"metadata\">\n{}\n</context>",
-        parts.join("\n")
-    ))
 }
 
 /// Format a user-friendly fallback message when the agent exceeds max tool steps
@@ -3131,402 +2974,6 @@ async fn persist_deadline_fallback(
         thinking: None,
         usage: None,
     })
-}
-
-/// Execute tool-use blocks from a response and push both assistant and
-/// tool-result messages onto the request. Returns summaries of each tool call
-/// for persistence in conversation metadata.
-#[allow(clippy::too_many_arguments)]
-async fn process_tool_calls(
-    response_content: Vec<LlmResponseContent>,
-    tools: &ToolRegistry,
-    skill_tools: &HashMap<String, &ResolvedSkillTool>,
-    skill_timeout: u64,
-    tool_ctx: &ToolContext<'_>,
-    request: &mut LlmRequest,
-    step: u32,
-    mcp_manager: Option<&McpManager>,
-    long_running_ctx: Option<&executor::LongRunningContext>,
-    db: &AsyncDatabase,
-    session_id: &str,
-    store_tool_calls: bool,
-    llm_call_id: Option<&str>,
-    // Send-message turn boundary state (#771). Passed from the agent loop to
-    // enable intra-step write gating: after a send_message call within this
-    // step, subsequent send_message calls are suppressed (conversation mode only).
-    send_message_boundary_active: &mut bool,
-    suppressed_write_tools: &mut Vec<String>,
-    send_message_text_capture: &mut String,
-    // Whether this is a conversation-mode run. The send_message boundary
-    // guard only applies in conversation mode — silent/callback modes use
-    // send_message as a notification mechanism, not a user-dialog interaction.
-    conversation_mode: bool,
-) -> Vec<ToolCallSummary> {
-    let mut tool_results: Vec<LlmContentBlock> = Vec::new();
-    let mut summaries = Vec::new();
-    let mut image_bytes_budget = crate::planning::policy::MAX_IMAGE_BYTES_PER_STEP;
-    // Per-turn dedup: if the LLM emits two tool_use blocks with identical
-    // (name, arguments) in a single response, execute the tool once and reuse
-    // the cached output for subsequent blocks. This defends the engine against
-    // provider-side tool_use duplication (see #582). Scope is strictly this
-    // function call — duplicates across steps or turns are unaffected.
-    let mut dedup_cache: HashMap<(String, String), ToolOutput> = HashMap::new();
-    for block in &response_content {
-        if let LlmResponseContent::ToolCall {
-            id,
-            name,
-            arguments,
-        } = block
-        {
-            // Send-message turn boundary gate (#771): intra-step suppression
-            // for duplicate send_message calls in conversation mode. Within
-            // the same LLM response, a second send_message after the first is
-            // suppressed (the agent should combine messages or wait for user
-            // input). Silent/callback modes are exempt — send_message is the
-            // notification mechanism there, not a user-dialog interaction.
-            // The inter-step gate (after process_tool_calls returns) prevents
-            // cross-step writes by forcing EndTurn.
-            if *send_message_boundary_active && name == "send_message" && conversation_mode {
-                let input_summary_for_suppressed = truncate_summary(
-                    &arguments.to_string(),
-                    crate::planning::policy::INPUT_SUMMARY_MAX,
-                );
-                let suppressed_msg =
-                    "[mika-engine] Tool call suppressed: send_message turn boundary (#771). \
-                     A second send_message is not permitted after the first in the same turn. \
-                     Combine your messages into a single send_message call."
-                        .to_string();
-                warn!(
-                    trace_id = %tool_ctx.trace_id,
-                    tool = %name,
-                    step,
-                    "send_message_turn_boundary: suppressed duplicate send_message"
-                );
-                summaries.push(ToolCallSummary {
-                    step,
-                    name: name.clone(),
-                    input_summary: scrub_secrets(&input_summary_for_suppressed).into_owned(),
-                    output_summary: truncate_summary(
-                        &suppressed_msg,
-                        crate::planning::policy::OUTPUT_SUMMARY_MAX,
-                    ),
-                    success: false,
-                    non_zero_exit: false,
-                });
-                suppressed_write_tools.push(name.clone());
-                // Return a tool_result so the conversation history stays paired
-                tool_results.push(LlmContentBlock::ToolResult {
-                    tool_call_id: id.clone(),
-                    content: LlmToolResultContent::Text(suppressed_msg),
-                    is_error: true,
-                });
-                continue;
-            }
-
-            let dedup_key = (
-                name.clone(),
-                serde_json::to_string(arguments).unwrap_or_default(),
-            );
-            let output = if let Some(cached) = dedup_cache.get(&dedup_key) {
-                warn!(
-                    trace_id = %tool_ctx.trace_id,
-                    tool = %name,
-                    step,
-                    cached_was_error = cached.is_error,
-                    "duplicate tool_use block suppressed; reusing prior result"
-                );
-                // Clone the cached result but strip images: the LLM already
-                // received them in the tool_result paired with the first
-                // duplicate's id, so re-emitting them would both waste the
-                // shared `image_bytes_budget` and send redundant bytes on
-                // the API request. Text content is preserved so the
-                // duplicate tool_use id still gets a meaningful pair.
-                let mut reused = cached.clone();
-                reused.images.clear();
-                reused
-            } else {
-                debug!(tool = %name, "executing tool");
-                let input_summary = scrub_secrets(&truncate_summary(
-                    &arguments.to_string(),
-                    crate::planning::policy::INPUT_SUMMARY_MAX,
-                ))
-                .into_owned();
-                let dispatch = ToolDispatchCtx {
-                    tools,
-                    skill_tools,
-                    ctx: tool_ctx,
-                    skill_timeout,
-                    mcp_manager,
-                    long_running_ctx,
-                };
-                let tool_start = std::time::Instant::now();
-                let output = execute_tool(&dispatch, name, arguments.clone()).await;
-                let tool_latency_ms = tool_start.elapsed().as_millis() as u64;
-
-                // Record full tool call in database
-                if store_tool_calls {
-                    let tool_id = uuid::Uuid::new_v4().to_string();
-                    let (tool_source, tool_skill_name) = if tools.get(name).is_some() {
-                        ("builtin", None)
-                    } else if let Some(st) = skill_tools.get(name.as_str()) {
-                        let sn = st
-                            .skill_dir
-                            .file_name()
-                            .and_then(|n| n.to_str())
-                            .unwrap_or("unknown")
-                            .to_string();
-                        ("skill", Some(sn))
-                    } else {
-                        ("mcp", None)
-                    };
-                    let non_zero = !output.is_error && has_non_zero_exit_prefix(&output.content);
-                    let input_json = serde_json::to_string(arguments).unwrap_or_default();
-                    let err_msg = if output.is_error {
-                        Some(output.content.as_str())
-                    } else {
-                        None
-                    };
-                    if let Err(e) = db
-                        .save_tool_call(
-                            &tool_id,
-                            session_id,
-                            Some(tool_ctx.trace_id),
-                            llm_call_id,
-                            step,
-                            name,
-                            tool_source,
-                            tool_skill_name.as_deref(),
-                            Some(&input_json),
-                            Some(&output.content),
-                            !output.is_error && !non_zero,
-                            non_zero,
-                            tool_latency_ms,
-                            err_msg,
-                        )
-                        .await
-                    {
-                        warn!(tool = %name, error = %e, "failed to save tool_call record");
-                    }
-                }
-
-                let image_count = output.images.len();
-                let output_summary = {
-                    let raw = if image_count > 0 {
-                        truncate_summary(
-                            &format!("{} [+{image_count} image(s)]", output.content),
-                            crate::planning::policy::OUTPUT_SUMMARY_MAX,
-                        )
-                    } else {
-                        truncate_summary(
-                            &output.content,
-                            crate::planning::policy::OUTPUT_SUMMARY_MAX,
-                        )
-                    };
-                    // Scrub secret-shaped values from metadata summaries (#908).
-                    scrub_secrets(&raw).into_owned()
-                };
-                let non_zero_exit = !output.is_error && has_non_zero_exit_prefix(&output.content);
-                let tool_succeeded = !output.is_error && !non_zero_exit;
-                summaries.push(ToolCallSummary {
-                    step,
-                    name: name.clone(),
-                    input_summary,
-                    output_summary,
-                    success: tool_succeeded,
-                    non_zero_exit,
-                });
-
-                // Send-message turn boundary activation (#771): after a
-                // successful send_message in conversation mode, mark the
-                // boundary as active. Silent/callback modes are exempt.
-                if name == "send_message" && tool_succeeded && conversation_mode {
-                    *send_message_boundary_active = true;
-                    // Capture the send_message text for structured logging.
-                    // Use the arguments (which contain the text parameter).
-                    if send_message_text_capture.is_empty()
-                        && let Some(text_val) = arguments.get("text").and_then(|v| v.as_str())
-                    {
-                        *send_message_text_capture = truncate_summary(text_val, 200);
-                    }
-                }
-
-                dedup_cache.insert(dedup_key, output.clone());
-                output
-            };
-
-            let image_count = output.images.len();
-            let content = if output.images.is_empty() {
-                LlmToolResultContent::Text(output.content)
-            } else {
-                let mut blocks = vec![LlmToolResultBlock::Text(output.content)];
-                let mut included = 0;
-                for img in output.images {
-                    let img_bytes = img.data.len();
-                    if img_bytes > image_bytes_budget {
-                        break;
-                    }
-                    image_bytes_budget -= img_bytes;
-                    included += 1;
-                    blocks.push(LlmToolResultBlock::Image(LlmImage {
-                        media_type: img.media_type,
-                        data: img.data,
-                    }));
-                }
-                if included < image_count {
-                    let skipped = image_count - included;
-                    blocks.push(LlmToolResultBlock::Text(format!(
-                        "[{skipped} image(s) skipped: step memory budget exceeded]"
-                    )));
-                    warn!(
-                        included,
-                        skipped, "image budget exceeded, skipped images in tool result"
-                    );
-                }
-                if included == 0 {
-                    // All images skipped — fall back to text-only
-                    LlmToolResultContent::Text(
-                        blocks
-                            .into_iter()
-                            .filter_map(|b| match b {
-                                LlmToolResultBlock::Text(text) => Some(text),
-                                _ => None,
-                            })
-                            .collect::<Vec<_>>()
-                            .join("\n"),
-                    )
-                } else {
-                    LlmToolResultContent::Blocks(blocks)
-                }
-            };
-            tool_results.push(LlmContentBlock::ToolResult {
-                tool_call_id: id.clone(),
-                content,
-                is_error: output.is_error,
-            });
-        }
-    }
-
-    // Convert response content to assistant message blocks
-    let assistant_blocks: Vec<LlmContentBlock> =
-        mika_common::llm::response_content_to_blocks(&response_content);
-    request.messages.push(LlmMessage {
-        role: LlmRole::Assistant,
-        content: LlmContent::Blocks(assistant_blocks),
-    });
-    request.messages.push(LlmMessage {
-        role: LlmRole::Tool,
-        content: LlmContent::Blocks(tool_results),
-    });
-    summaries
-}
-
-/// Resources for tool dispatch, bundled to reduce argument count.
-struct ToolDispatchCtx<'a> {
-    tools: &'a ToolRegistry,
-    skill_tools: &'a HashMap<String, &'a ResolvedSkillTool>,
-    ctx: &'a ToolContext<'a>,
-    skill_timeout: u64,
-    mcp_manager: Option<&'a McpManager>,
-    long_running_ctx: Option<&'a executor::LongRunningContext>,
-}
-
-/// Execute a single tool with timeout.
-///
-/// Routing: builtin tools (from ToolRegistry) first, then skill-defined tools,
-/// then "unknown tool" error.
-async fn execute_tool(
-    dispatch: &ToolDispatchCtx<'_>,
-    name: &str,
-    input: serde_json::Value,
-) -> ToolOutput {
-    debug!(tool = %name, "tool_execution");
-
-    // Pre-compute input excerpt for timeout diagnostics (#900). Must happen
-    // before `input` is moved into the execute call.
-    let input_excerpt: String = serde_json::to_string(&input)
-        .unwrap_or_default()
-        .chars()
-        .take(TOOL_TIMEOUT_INPUT_EXCERPT_LEN)
-        .collect();
-
-    // 1. Try builtin tool
-    if let Some(tool) = dispatch.tools.get(name) {
-        let timeout = tool
-            .timeout_secs()
-            .unwrap_or(crate::planning::policy::TOOL_TIMEOUT_SECS);
-        return match tokio::time::timeout(
-            std::time::Duration::from_secs(timeout),
-            tool.execute(input, dispatch.ctx),
-        )
-        .await
-        {
-            Ok(Ok(output)) => output,
-            Ok(Err(e)) => {
-                warn!(tool = %name, error = %e, "tool execution failed");
-                ToolOutput::error(format!("Tool error: {e}"))
-            }
-            Err(_) => {
-                warn!(tool = %name, timeout_secs = timeout, input_excerpt = %input_excerpt, "tool execution timed out");
-                ToolOutput::error(format!("Tool '{name}' timed out after {timeout}s"))
-            }
-        };
-    }
-
-    // 2. Try skill-defined tool
-    if let Some(skill_tool) = dispatch.skill_tools.get(name) {
-        // Builtin skill handlers dispatch to Rust functions with ToolContext access
-        if let ToolHandler::Builtin { function } = &skill_tool.handler {
-            let timeout = dispatch.skill_timeout;
-            return match tokio::time::timeout(
-                std::time::Duration::from_secs(timeout),
-                builtin_handlers::execute(function, input, dispatch.ctx),
-            )
-            .await
-            {
-                Ok(output) => output,
-                Err(_) => {
-                    warn!(tool = %name, timeout_secs = timeout, input_excerpt = %input_excerpt, "builtin handler timed out");
-                    ToolOutput::error(format!("Builtin tool '{name}' timed out after {timeout}s"))
-                }
-            };
-        }
-        return executor::execute_skill_tool(
-            skill_tool,
-            input,
-            dispatch.skill_timeout,
-            dispatch.long_running_ctx,
-            dispatch.ctx.github_token,
-            dispatch.ctx.callback_task_id,
-            if dispatch.ctx.callback_task_id.is_some() {
-                Some(dispatch.ctx.db)
-            } else {
-                None
-            },
-        )
-        .await;
-    }
-
-    // 3. Try MCP tool (external server)
-    if let Some(mcp) = dispatch.mcp_manager
-        && mcp.is_mcp_tool(name)
-    {
-        return match tokio::time::timeout(
-            std::time::Duration::from_secs(crate::planning::policy::TOOL_TIMEOUT_SECS),
-            mcp.call_tool(name, input),
-        )
-        .await
-        {
-            Ok(output) => output,
-            Err(_) => {
-                warn!(tool = %name, "MCP tool execution timed out");
-                let timeout_secs = crate::planning::policy::TOOL_TIMEOUT_SECS;
-                ToolOutput::error(format!("MCP tool '{name}' timed out after {timeout_secs}s"))
-            }
-        };
-    }
-
-    warn!(tool = %name, "unknown tool requested");
-    ToolOutput::error(format!("Unknown tool: {name}"))
 }
 
 // -- Summary Gating (Axis 4 + Axis 3) --
@@ -7098,343 +6545,6 @@ mod tests {
         assert_eq!(max_skill_timeout(&matched, "groq", "llama"), 30);
     }
 
-    // -- ToolCallSummary and metadata tests --
-
-    #[test]
-    fn test_truncate_summary_no_op_for_short_strings() {
-        assert_eq!(truncate_summary("hello", 10), "hello");
-    }
-
-    #[test]
-    fn test_truncate_summary_truncates_long_strings() {
-        let long = "a".repeat(300);
-        let result = truncate_summary(&long, 200);
-        assert!(result.len() <= 200);
-        assert!(result.ends_with("..."));
-    }
-
-    #[test]
-    fn test_truncate_summary_exact_length_not_truncated() {
-        let exact = "a".repeat(200);
-        assert_eq!(truncate_summary(&exact, 200), exact);
-    }
-
-    #[test]
-    fn test_truncate_summary_safe_with_multibyte_chars() {
-        // Euro sign is 3 bytes: \xe2\x82\xac
-        let s = "\u{20AC}".repeat(100); // 300 bytes, 100 chars
-        let result = truncate_summary(&s, 10);
-        assert!(result.ends_with("..."));
-        // Must not panic and must be valid UTF-8
-        assert!(std::str::from_utf8(result.as_bytes()).is_ok());
-    }
-
-    #[test]
-    fn test_truncate_summary_safe_with_emoji() {
-        // Emoji is 4 bytes
-        let s = "Hello \u{1F600} world! More text here to exceed the limit easily.";
-        let result = truncate_summary(s, 10);
-        assert!(result.ends_with("..."));
-        assert!(result.len() <= 13); // 10 bytes + "..."
-    }
-
-    #[test]
-    fn test_tool_calls_metadata_json_empty_returns_none() {
-        assert!(tool_calls_metadata_json(&[]).is_none());
-    }
-
-    #[test]
-    fn test_tool_calls_metadata_json_single_call() {
-        let summaries = vec![ToolCallSummary {
-            step: 0,
-            name: "search_memory".to_string(),
-            input_summary: r#"{"query":"meetings"}"#.to_string(),
-            output_summary: "Found 3 results".to_string(),
-            success: true,
-            non_zero_exit: false,
-        }];
-        let json = tool_calls_metadata_json(&summaries).unwrap();
-        let parsed: serde_json::Value = serde_json::from_str(&json).unwrap();
-        let calls = parsed["tool_calls"].as_array().unwrap();
-        assert_eq!(calls.len(), 1);
-        assert_eq!(calls[0]["name"], "search_memory");
-        assert_eq!(calls[0]["success"], true);
-    }
-
-    #[test]
-    fn test_tool_calls_metadata_json_respects_max_size() {
-        // Create many tool calls with large outputs to exceed crate::planning::policy::TOOL_METADATA_MAX
-        let summaries: Vec<ToolCallSummary> = (0..50)
-            .map(|i| ToolCallSummary {
-                step: i,
-                name: format!("tool_{i}"),
-                input_summary: "x".repeat(200),
-                output_summary: "y".repeat(300),
-                success: true,
-                non_zero_exit: false,
-            })
-            .collect();
-        let json = tool_calls_metadata_json(&summaries).unwrap();
-        // Must produce valid JSON within the size cap
-        assert!(
-            json.len() <= crate::planning::policy::TOOL_METADATA_MAX,
-            "metadata exceeded crate::planning::policy::TOOL_METADATA_MAX: {} chars",
-            json.len()
-        );
-        let parsed: serde_json::Value = serde_json::from_str(&json).unwrap();
-        assert!(!parsed["tool_calls"].as_array().unwrap().is_empty());
-    }
-
-    #[test]
-    fn test_tool_call_summary_truncates_large_inputs() {
-        // Simulate what happens when building a ToolCallSummary with large content
-        let large_input = "x".repeat(10_000);
-        let large_output = "y".repeat(10_000);
-        let input_summary =
-            truncate_summary(&large_input, crate::planning::policy::INPUT_SUMMARY_MAX);
-        let output_summary =
-            truncate_summary(&large_output, crate::planning::policy::OUTPUT_SUMMARY_MAX);
-
-        assert!(
-            input_summary.len() <= crate::planning::policy::INPUT_SUMMARY_MAX,
-            "input_summary too long: {} chars",
-            input_summary.len()
-        );
-        assert!(
-            output_summary.len() <= crate::planning::policy::OUTPUT_SUMMARY_MAX,
-            "output_summary too long: {} chars",
-            output_summary.len()
-        );
-        assert!(input_summary.ends_with("..."));
-        assert!(output_summary.ends_with("..."));
-    }
-
-    #[test]
-    fn test_all_entries_preserved_at_max_steps() {
-        // With reduced per-field limits, 10 entries with typical tool names should
-        // all fit within crate::planning::policy::TOOL_METADATA_MAX without tail-drop
-        let summaries: Vec<ToolCallSummary> = (0..10)
-            .map(|i| ToolCallSummary {
-                step: i,
-                name: "search_memory".to_string(),
-                input_summary: truncate_summary(
-                    &"x".repeat(10_000),
-                    crate::planning::policy::INPUT_SUMMARY_MAX,
-                ),
-                output_summary: truncate_summary(
-                    &"y".repeat(10_000),
-                    crate::planning::policy::OUTPUT_SUMMARY_MAX,
-                ),
-                success: true,
-                non_zero_exit: false,
-            })
-            .collect();
-        let json = tool_calls_metadata_json(&summaries).unwrap();
-        assert!(
-            json.len() <= crate::planning::policy::TOOL_METADATA_MAX,
-            "truncated summaries exceed cap: {} chars",
-            json.len()
-        );
-        let parsed: serde_json::Value = serde_json::from_str(&json).unwrap();
-        assert_eq!(
-            parsed["tool_calls"].as_array().unwrap().len(),
-            10,
-            "all 10 entries must be preserved"
-        );
-    }
-
-    #[test]
-    fn test_safety_net_drops_tail_on_overflow() {
-        // With pathologically long tool names or extreme content, the safety net
-        // tail-drop should still produce valid JSON within the cap
-        let summaries: Vec<ToolCallSummary> = (0..20)
-            .map(|i| ToolCallSummary {
-                step: i,
-                name: format!("mcp__very_long_server_name__tool_with_long_name_{i}"),
-                input_summary: "x".repeat(crate::planning::policy::INPUT_SUMMARY_MAX),
-                output_summary: "y".repeat(crate::planning::policy::OUTPUT_SUMMARY_MAX),
-                success: true,
-                non_zero_exit: false,
-            })
-            .collect();
-        let json = tool_calls_metadata_json(&summaries).unwrap();
-        assert!(
-            json.len() <= crate::planning::policy::TOOL_METADATA_MAX,
-            "safety net failed: {} chars",
-            json.len()
-        );
-        let parsed: serde_json::Value = serde_json::from_str(&json).unwrap();
-        let entries = parsed["tool_calls"].as_array().unwrap();
-        assert!(!entries.is_empty(), "must retain at least one entry");
-        assert!(entries.len() < 20, "some entries should have been dropped");
-    }
-
-    /// Regression test for #744: milestone-workflow turns with 21+ tool calls have
-    /// their tail entries dropped by the 4KB metadata cap. This is acceptable because
-    /// the dashboard now fetches from the `tool_calls` table (via `useTraceToolCalls`)
-    /// instead of parsing this metadata. The metadata path is only used for the LLM
-    /// history builder's `format_tool_summary_block()`.
-    #[test]
-    fn test_metadata_cap_drops_tail_on_milestone_workflow_turns() {
-        // Simulate a milestone-workflow turn: 14 bookkeeping calls + 7 status updates + 1 dispatch
-        // Use crate::planning::policy::INPUT_SUMMARY_MAX/crate::planning::policy::OUTPUT_SUMMARY_MAX length strings to guarantee the 4KB cap is hit
-        let tool_names = [
-            "run_gh",
-            "create_task",
-            "run_gh",
-            "resolve_issue_order",
-            "run_gh",
-            "list_tasks",
-            "list_tasks",
-            "create_task",
-            "create_task",
-            "create_task",
-            "create_task",
-            "create_task",
-            "create_task",
-            "update_task_status",
-            "update_task_status",
-            "update_task_status",
-            "update_task_status",
-            "update_task_status",
-            "update_task_status",
-            "update_task_status",
-            "run_claude_pilot",
-        ];
-        let summaries: Vec<ToolCallSummary> = tool_names
-            .iter()
-            .enumerate()
-            .map(|(i, name)| ToolCallSummary {
-                step: i as u32 / 3, // group into steps like the real agent loop
-                name: name.to_string(),
-                input_summary: truncate_summary(
-                    &"x".repeat(10_000),
-                    crate::planning::policy::INPUT_SUMMARY_MAX,
-                ),
-                output_summary: truncate_summary(
-                    &"y".repeat(10_000),
-                    crate::planning::policy::OUTPUT_SUMMARY_MAX,
-                ),
-                success: true,
-                non_zero_exit: false,
-            })
-            .collect();
-
-        assert_eq!(
-            summaries.len(),
-            21,
-            "milestone-workflow turn should have 21 calls"
-        );
-
-        let json = tool_calls_metadata_json(&summaries).unwrap();
-        assert!(
-            json.len() <= crate::planning::policy::TOOL_METADATA_MAX,
-            "metadata must respect cap: {} chars",
-            json.len()
-        );
-
-        let parsed: serde_json::Value = serde_json::from_str(&json).unwrap();
-        let entries = parsed["tool_calls"].as_array().unwrap();
-
-        // With max-length fields, the 4KB cap forces tail-drop — fewer entries than input.
-        // This is the documented limitation (#744). The dashboard uses the tool_calls table
-        // instead of this metadata, so tail-drop is acceptable for the LLM history context.
-        assert!(!entries.is_empty(), "must retain at least one entry");
-        assert!(
-            entries.len() < summaries.len(),
-            "4KB cap should force tail-drop: got {} entries from {} inputs",
-            entries.len(),
-            summaries.len()
-        );
-
-        // Verify structural integrity of kept entries
-        for entry in entries {
-            assert!(entry["name"].is_string(), "entries must have name");
-            assert!(entry["step"].is_number(), "entries must have step");
-        }
-    }
-
-    #[test]
-    fn test_format_tool_summary_block_valid_json() {
-        let json = r#"{"tool_calls":[{"step":0,"name":"tmux_send_command","input_summary":"{\"session\":\"mika\",\"text\":\"cargo test\"}","output_summary":"Command sent","success":true}]}"#;
-        let block = format_tool_summary_block(json).unwrap();
-        assert!(block.contains("tmux_send_command"));
-        assert!(block.contains("cargo test")); // input is now surfaced
-        assert!(block.contains("Command sent"));
-        assert!(block.starts_with("\n<context type=\"tool_history\""));
-        assert!(block.contains("</context>"));
-    }
-
-    #[test]
-    fn test_format_tool_summary_block_failed_tool() {
-        let json = r#"{"tool_calls":[{"step":0,"name":"bad_tool","input_summary":"","output_summary":"error","success":false}]}"#;
-        let block = format_tool_summary_block(json).unwrap();
-        assert!(block.contains("[FAILED]"));
-    }
-
-    #[test]
-    fn test_format_tool_summary_block_skips_malformed_entries() {
-        // One good entry, one missing name — should produce partial result
-        let json = r#"{"tool_calls":[{"step":0,"name":"good_tool","input_summary":"","output_summary":"ok","success":true},{"step":1,"output_summary":"no name"}]}"#;
-        let block = format_tool_summary_block(json).unwrap();
-        assert!(block.contains("good_tool"));
-        // The malformed entry should be skipped, not cause None
-    }
-
-    #[test]
-    fn test_format_tool_summary_block_empty_calls_returns_none() {
-        let json = r#"{"tool_calls":[]}"#;
-        assert!(format_tool_summary_block(json).is_none());
-    }
-
-    #[test]
-    fn test_format_tool_summary_block_invalid_json_returns_none() {
-        assert!(format_tool_summary_block("not json").is_none());
-    }
-
-    #[test]
-    fn test_format_tool_summary_block_non_zero_exit_old_format() {
-        // Backward compat: old metadata had success: true with non_zero_exit: true
-        let json = r#"{"tool_calls":[{"step":0,"name":"shell_exec","input_summary":"grep foo","output_summary":"Exit code: 1\nno matches","success":true,"non_zero_exit":true}]}"#;
-        let block = format_tool_summary_block(json).unwrap();
-        assert!(
-            block.contains("[NON-ZERO]"),
-            "expected [NON-ZERO] tag in: {block}"
-        );
-        assert!(!block.contains("[FAILED]"));
-    }
-
-    #[test]
-    fn test_format_tool_summary_block_non_zero_exit_new_format() {
-        // New format: success is false when non_zero_exit is true
-        let json = r#"{"tool_calls":[{"step":0,"name":"shell_exec","input_summary":"grep foo","output_summary":"Exit code: 1\nno matches","success":false,"non_zero_exit":true}]}"#;
-        let block = format_tool_summary_block(json).unwrap();
-        assert!(
-            block.contains("[NON-ZERO]"),
-            "expected [NON-ZERO] tag in: {block}"
-        );
-        assert!(!block.contains("[FAILED]"));
-    }
-
-    #[test]
-    fn test_format_tool_summary_block_non_zero_exit_missing_defaults_false() {
-        // Backward compat: old metadata without non_zero_exit field
-        let json = r#"{"tool_calls":[{"step":0,"name":"shell_exec","input_summary":"ls","output_summary":"files","success":true}]}"#;
-        let block = format_tool_summary_block(json).unwrap();
-        assert!(!block.contains("[NON-ZERO]"));
-        assert!(!block.contains("[FAILED]"));
-    }
-
-    #[test]
-    fn test_has_non_zero_exit_prefix() {
-        assert!(has_non_zero_exit_prefix("Exit code: 1\nsome output"));
-        assert!(has_non_zero_exit_prefix("Exit code: 127\n"));
-        assert!(has_non_zero_exit_prefix("Killed by signal: 9\n"));
-        assert!(!has_non_zero_exit_prefix("Exit code: unknown\nstuff"));
-        assert!(!has_non_zero_exit_prefix("All good, no errors"));
-        assert!(!has_non_zero_exit_prefix(""));
-    }
-
     #[test]
     fn test_format_step_exceeded_fallback_non_zero_exit_old_format() {
         // Backward compat: old metadata had success: true with non_zero_exit: true
@@ -8702,7 +7812,6 @@ mod tests {
 
     #[test]
     fn test_deferred_dispatch_satisfied_requires_run_claude_pilot() {
-        use crate::agent::ToolCallSummary;
         // No tools → not satisfied
         assert!(!deferred_dispatch_satisfied(&[]));
 
@@ -9475,7 +8584,10 @@ mod tests {
             long_org
         );
         // Truncate to crate::planning::policy::INPUT_SUMMARY_MAX to simulate what ToolCallSummary does.
-        let truncated = truncate_summary(&input, crate::planning::policy::INPUT_SUMMARY_MAX);
+        let truncated = crate::tool_execution::truncate_summary(
+            &input,
+            crate::planning::policy::INPUT_SUMMARY_MAX,
+        );
         let summaries = vec![run_gh_summary(&truncated)];
 
         let result = detect_milestone_close_claim_without_patch(
